Refactor external script handling to use classes for all supported script types
